Starting code re-factor, breaking out easy bits
