+++ llama-server/Dockerfile
@@ -29,4 +29,4 @@ ENV \
 EXPOSE 80
 
 # command to run on container start
-CMD [ "python", "./app.py" ]
+CMD [ "python", "./main.py" ]
