+++ llama-server/src/common/constants.py
@@ -1,0 +1,6 @@
+'''
+Contains constant variables within the application
+'''
+
+default_group       = "default"
+default_keepalive   = 86400
