+++ llama-server/src/common/environment.py
@@ -1,0 +1,5 @@
+'''
+Contains variables from OS environment
+'''
+
+from os import environ
