+++ llama-server/src/config.yml
@@ -1,14 +1,15 @@
-### Web Server Settings ###
-
+#############################
+###  Web Server Settings  ###
+#############################
 # Port webserver listens on. Defaults to 5000 if not set.
 port: 80
 
 # Server IP to listen for web traffic, 0.0.0.0 is all available IP's. Defaults to 127.0.0.1 if not set.
 host: 0.0.0.0
 
-
+#############################
 ### Registration Settings ###
-
+#############################
 # How long before service discovery removes an enpoint that has not
 # checked in. This value is in seconds and is set by default to the 
 # value listed below. Probe settings will overwrite this value.
@@ -18,13 +19,14 @@ keepalive: 86400
 # settings will overwrite this value. Defaults to 'none' if not set.
 group: default
 
-
-### LLAMA Settings ###
+#############################
+###    LLAMA Settings     ###
+#############################
 # How often should the scraper pull data and collector summarize in seconds.
 interval: 10
 
-
+#############################
 ### Application Debugging ###
-
+#############################
 # Debug logging level, True or False
 verbose: False
